Fix color bleed
