+++ demos/storybook/stories/user-menu/with-custom-colors.tsx
@@ -4,7 +4,7 @@ import * as Colors from '@pxblue/colors';
 import { UserMenu } from '@pxblue/react-components';
 import { color } from '@storybook/addon-knobs';
 import { StoryFnReactReturnType } from '@storybook/react/dist/client/preview/types';
-import { menuItems } from './with-default-colors';
+import { getMenu } from './with-basic-usage';
 
 export const withCustomColors = (): StoryFnReactReturnType => {
     const useStyles = makeStyles({
@@ -18,7 +18,7 @@ export const withCustomColors = (): StoryFnReactReturnType => {
     });
     const classes = useStyles();
     const avatar = <Avatar classes={{ root: classes.root }}>CD</Avatar>;
-    const items = menuItems[0];
+    const items = getMenu()[0];
     items.fontColor = color('fontColor', Colors.gray[500], 'Menu');
     items.iconColor = color('iconColor', Colors.blue[800], 'Menu');
