+++ demos/storybook/stories/user-menu/with-menu-placement-options.tsx
@@ -3,12 +3,10 @@ import { UserMenu } from '@pxblue/react-components';
 import { select } from '@storybook/addon-knobs';
 import { StoryFnReactReturnType } from '@storybook/react/dist/client/preview/types';
 import React from 'react';
-import { menuItems } from './with-default-colors';
+import { getMenu } from './with-basic-usage';
 
 export const withMenuPlacementOptions = (): StoryFnReactReturnType => {
     const avatar = <Avatar>CD</Avatar>;
-    const items = menuItems[0];
-
     const anchorOriginHorizontal = select('anchorOrigin.horizontal', ['left', 'center', 'right'], 'left');
     const anchorOriginVertical = select('anchorOrigin.vertical', ['top', 'center', 'bottom'], 'top');
     const transformOriginHorizontal = select('transformOrigin.horizontal', ['left', 'center', 'right'], 'left');
@@ -17,7 +15,7 @@ export const withMenuPlacementOptions = (): StoryFnReactReturnType => {
     return (
         <UserMenu
             avatar={avatar}
-            menuGroups={[items]}
+            menuGroups={getMenu()}
             MenuProps={{
                 anchorOrigin: { horizontal: anchorOriginHorizontal, vertical: anchorOriginVertical },
                 transformOrigin: { horizontal: transformOriginHorizontal, vertical: transformOriginVertical },
